Backend_Damir/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-cleanup==6.0.0
django-filter==21.1
djangorestframework==3.13.1
Pillow==9.0.1
pytz==2022.1
sqlparse==0.4.2
